Add docstrings for important classes
