+++ ogc-example/src/main/scala/geotrellis/server/ogc/conf/StyleConf.scala
@@ -4,12 +4,14 @@ import geotrellis.server.ogc._
 
 import geotrellis.raster.render.{ColorMap, ColorRamp}
 
+/** The trait implemented by different style configuration options */
 sealed trait StyleConf {
   def name: String
   def title: String
   def toStyle: OgcStyle
 }
 
+/** Styling in which a color scheme is known but not the values that these colors should map to */
 final case class ColorRampConf(
   name: String,
   title: String,
@@ -20,6 +22,7 @@ final case class ColorRampConf(
   def toStyle: OgcStyle = ColorRampStyle(name, title, colors, stops, legends)
 }
 
+/** Styling in which both a color scheme and the data-to-color mapping is known */
 final case class ColorMapConf(
   name: String,
   title: String,
