+++ ogc/src/main/scala/geotrellis/server/ogc/wms/WmsParentLayerMeta.scala
@@ -2,6 +2,7 @@ package geotrellis.server.ogc.wms
 
 import geotrellis.proj4.CRS
 
+/** Parent layer metadata class (used in configuration and reporting capabilities) */
 case class WmsParentLayerMeta(
   name: Option[String],
   title: String,
