+++ ogc/src/main/scala/geotrellis/server/ogc/wmts/GeotrellisTileMatrixSet.scala
@@ -8,6 +8,7 @@ import opengis.wmts.TileMatrixSet
 
 import java.net.URI
 
+/** A collection of tile matrices; most commonly forming a pyramid of different resolutions */
 case class GeotrellisTileMatrixSet(
   identifier: String,
   supportedCrs: CRS,
