+++ ogc/src/main/scala/geotrellis/server/ogc/wmts/WmtsModel.scala
@@ -4,6 +4,7 @@ import geotrellis.server.ogc._
 import geotrellis.spark.tiling._
 import geotrellis.proj4._
 
+/** This class holds all the information necessary to construct a response to a WMTS request */
 case class WmtsModel(
   serviceMetadata: ows.ServiceMetadata,
   matrices: List[GeotrellisTileMatrixSet],
